requirements.txt
werkzeug==2.0.3
Flask==2.0.2
mysql-connector-python==8.0.32
python-decouple==3.8
pandas>=2.0.0
numpy>=1.24.0
xarray>=0.21.1
scipy>=1.10.0
netCDF4>=1.6.5
urllib3==2.0.3
mpld3==0.5.9
pysd>=3.14.0
pyngrok==6.0.0